requirements.txt
ArchivesSnake==0.9.1
psutil==5.8.0
pytest==6.2.4
vcrpy==4.1.1
